+++ code/tools/ci/psm1/cfxUploadServerSymbols.psm1
@@ -53,20 +53,24 @@ function Invoke-UploadServerSymbols {
 
     # chdir to the directory to avoid converting path to what rsync would expect
     Push-Location $symUploadDir
-        # remember current PATH content
-        $oldEnvPath = $env:PATH
-
-        # prepend PATH with MSYS2 bin dir so it can use it's own SSH
-        $msys2UsrBin = [IO.Path]::Combine($Tools.getMSYS2Root(), "usr\bin");
-        $env:PATH = "$msys2UsrBin;$env:PATH"
-
-        & $rsync -r -a -v -e $env:RSH_SYMBOLS_COMMAND ./ $env:SSH_SYMBOLS_TARGET
-        if ($LASTEXITCODE -ne 0) {
-            $Context.addBuildWarning("Failed to upload symbols, rsync failed")
+        if ($Context.IsDryRun) {
+            Write-Output "DRY RUN: Would upload symbols"
+        } else {
+            # remember current PATH content
+            $oldEnvPath = $env:PATH
+    
+            # prepend PATH with MSYS2 bin dir so it can use it's own SSH
+            $msys2UsrBin = [IO.Path]::Combine($Tools.getMSYS2Root(), "usr\bin");
+            $env:PATH = "$msys2UsrBin;$env:PATH"
+    
+            & $rsync -r -a -v -e $env:RSH_SYMBOLS_COMMAND ./ $env:SSH_SYMBOLS_TARGET
+            if ($LASTEXITCODE -ne 0) {
+                $Context.addBuildWarning("Failed to upload symbols, rsync failed")
+            }
+    
+            # restore PATH
+            $env:PATH = $oldEnvPath
         }
-
-        # restore PATH
-        $env:PATH = $oldEnvPath
     Pop-Location
 
     if ($Context.IsReleaseBuild) {
